+++ tests/idea/laravelInsight/fluent/PropertyWithoutAnnotationInspectionTest.java
@@ -1,0 +1,11 @@
+package net.rentalhost.idea.laravelInsight.fluent;
+
+import net.rentalhost.suite.FixtureSuite;
+
+public class PropertyWithoutAnnotationInspectionTest extends FixtureSuite {
+    public void testAll() {
+        inspectTool(PropertyWithoutAnnotationInspection.class)
+            .addTestFile("laravelInsight/fluent/PropertyWithoutAnnotationInspection.php")
+            .highlightTest();
+    }
+}